feat: extract useSiteThemeCssVars custom hook
